requirements.txt
Click
Mako
Beaker
MarkupSafe
requests
PyYAML
Twisted>=15.0.0
tablib
cyclone
six
pycrypto
MySQL-python
SQLAlchemy
pyzmq
txzmq
msgpack-python
psutil
txradius
redis
supervisor